# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: CervixDetectAI API
  version: 1.0.0
  description: OpenAPI specification for CervixDetectAI backend.
servers:
  - url: /api
    description: Current host (recommended for production)
  - url: https://{domain}/api
    description: Custom domain
    variables:
      domain:
        default: api.example.com
  - url: http://localhost:4000/api
    description: Local dev (4000)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
      required:
        - success
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer
    UserPublic:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          nullable: true
        real_name:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        hospital_id:
          type: string
          nullable: true
        employee_id:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        role:
          type: string
          enum: [admin, doctor, user]
        status:
          type: string
          enum: [active, disabled]
    User:
      allOf:
        - $ref: '#/components/schemas/UserPublic'
        - type: object
          properties:
            subscription_type:
              type: string
              enum: [none, monthly, yearly, package]
            subscription_expires_at:
              type: string
              format: date-time
              nullable: true
            remaining_credits:
              type: integer
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
...
    ReportResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                report:
                  $ref: '#/components/schemas/ReportMock'
...
    ReportMock:
      type: object
      properties:
        id:
          type: string
        patientName:
          type: string
        age:
          type: integer
        patientId:
          type: string
        date:
          type: string
        riskLevel:
          type: string
        status:
          type: string
        confidence:
          type: number
        recommendation:
          type: string
...
    UserResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                user:
                  $ref: '#/components/schemas/User'
    UserListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                users:
                  type: array
                  items:
                    $ref: '#/components/schemas/User'
                pagination:
                  $ref: '#/components/schemas/Pagination'
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    put:
      tags: [Users]
      summary: Update current user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                real_name:
                  type: string
                phone:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /users/me/password:
    put:
      tags: [Users]
      summary: Change password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
  /users/me/avatar:
    post:
      tags: [Users]
      summary: Upload avatar
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [avatar]
              properties:
                avatar:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /users:
    get:
      tags: [Users]
      summary: List users (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: role
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    put:
      tags: [Users]
      summary: Update user (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                real_name:
                  type: string
                phone:
                  type: string
                role:
                  type: string
                status:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    delete:
      tags: [Users]
      summary: Delete user (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /settings/users:
    get:
      tags: [Settings]
      summary: List settings users (mock)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    post:
      tags: [Settings]
      summary: Create settings user (mock)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...